src/service/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.0.1
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.3
MarkupSafe==2.1.5
orjson==3.10.3
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.7.0
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
PyYAML==6.0.1
redis==5.0.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
text-unidecode==1.3
typing_extensions==4.11.0
ujson==5.9.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
